+++ src/vmm/src/devices/virtio/vsock/unix/muxer_rxq.rs
@@ -20,8 +20,8 @@ use std::collections::{HashMap, VecDeque};
 use super::super::VsockChannel;
 use super::defs;
 use super::muxer::{ConnMapKey, MuxerRx};
-use crate::devices::virtio::vsock::csm::{VsockConnection, VsockConnectionBackend};
-use crate::devices::virtio::vsock::unix::MuxerConn;
+use crate::devices::virtio::vsock::csm::VsockConnection;
+use crate::devices::virtio::vsock::unix::ConnBackend;
 
 /// The muxer RX queue.
 #[derive(Debug)]
@@ -47,7 +47,7 @@ impl MuxerRxQ {
     /// Note: the resulting queue may still be desynchronized, if there are too many connections
     ///       that have pending RX data. In that case, the muxer will first drain this queue, and
     ///       then try again to build a synchronized one.
-    pub fn from_conn_map(conn_map: &HashMap<ConnMapKey, MuxerConn>) -> Self {
+    pub fn from_conn_map(conn_map: &HashMap<ConnMapKey, VsockConnection<ConnBackend>>) -> Self {
         let mut q = VecDeque::new();
         let mut synced = true;